# src/docs/swagger.yaml
basePath: /api
definitions:
...
  models.AddBotFriendRequest:
    properties:
      alias:
        type: string
      app_key:
        type: string
      bot_config_id:
        type: integer
      priority:
        type: integer
    required:
    - app_key
    - bot_config_id
    type: object
...
  models.CreateBotConfigRequest:
    properties:
      config_name:
        type: string
      description:
        type: string
      function_name:
        minLength: 1
        type: string
      max_tokens:
        type: integer
      mcp_server_url:
        type: string
      model_id:
        type: integer
      parameters:
        additionalProperties: true
        type: object
      temperature:
        type: number
      visibility:
        description: private/public
        type: string
    required:
    - config_name
    - description
    - function_name
    - model_id
    type: object
...
  models.UpdateBotConfigRequest:
    properties:
      config_name:
        type: string
      description:
        type: string
      function_name:
        type: string
      max_tokens:
        type: integer
      mcp_server_url:
        type: string
      parameters:
        additionalProperties: true
        type: object
      temperature:
        type: number
      visibility:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: 怒喵 API 文档
  version: "1.0"
paths:
...
  /api/app/friends/bots:
    get:
      consumes:
      - application/json
      description: 获取用户的所有Bot好友
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取Bot好友列表
      tags:
      - 用户好友管理
    post:
      consumes:
      - application/json
      description: 添加Bot为好友
      parameters:
      - description: 好友信息
        in: body
        name: friend
        required: true
        schema:
          $ref: '#/definitions/models.AddBotFriendRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 添加成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 添加Bot好友
      tags:
      - 用户好友管理
  /api/app/friends/bots/{bot_config_id}:
    delete:
      consumes:
      - application/json
      description: 删除Bot好友关系
      parameters:
      - description: Bot配置ID
        in: path
        name: bot_config_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 好友不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除Bot好友
      tags:
      - 用户好友管理
...
  /api/v2/bots:
    post:
      consumes:
      - application/json
      description: 创建新的Bot配置
      parameters:
      - description: 配置信息
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.CreateBotConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建Bot配置
      tags:
      - Bot配置管理
  /api/v2/bots/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的Bot配置
      parameters:
      - description: Bot配置ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 配置不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除Bot配置
      tags:
      - Bot配置管理
    get:
      consumes:
      - application/json
      description: 根据ID获取Bot配置详情
      parameters:
      - description: Bot配置ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 配置不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取Bot配置详情
      tags:
      - Bot配置管理
    put:
      consumes:
      - application/json
      description: 更新指定的Bot配置
      parameters:
      - description: Bot配置ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBotConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 配置不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新Bot配置
      tags:
      - Bot配置管理
  /api/v2/bots/my:
    get:
      consumes:
      - application/json
      description: 获取当前用户创建的所有Bot配置
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取我创建的Bot列表
      tags:
      - Bot配置管理
  /api/v2/bots/search:
    get:
      consumes:
      - application/json
      description: 搜索Bot配置，支持bot_hash精确搜索和bot_name模糊搜索
      parameters:
      - description: 搜索关键词
        in: query
        name: query
        required: true
        type: string
      - description: 搜索类型：hash/name/description，默认为name
        in: query
        name: search_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 搜索Bot配置
      tags:
      - Bot配置管理
...
swagger: "2.0"